feat: thinking screen basic UI
