+++ lib/home/widgets/thinking_view.dart
@@ -1,3 +1,4 @@
+import 'package:app_ui/app_ui.dart';
 import 'package:dash_ai_search/home/home.dart';
 import 'package:dash_ai_search/l10n/l10n.dart';
 import 'package:flutter/material.dart';
@@ -8,34 +9,88 @@ class ThinkingView extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final theme = Theme.of(context);
-    final l10n = context.l10n;
-
-    final state = context.watch<HomeBloc>().state;
-    final isAnimating = state.status == Status.askQuestionToThinking;
+    final isAnimating = context.select(
+      (HomeBloc bloc) => bloc.state.status == Status.askQuestionToThinking,
+    );
     final query = context.select((HomeBloc bloc) => bloc.state.query);
     return AnimatedOpacity(
       opacity: isAnimating ? 1 : 0,
       duration: const Duration(milliseconds: 500),
-      child: Center(
-        child: SingleChildScrollView(
-          child: Column(
-            mainAxisAlignment: MainAxisAlignment.center,
-            children: [
-              Text(
-                l10n.thinkingHeadline,
-                textAlign: TextAlign.center,
-                style: theme.textTheme.bodyMedium,
-              ),
-              Text(
-                query,
-                textAlign: TextAlign.center,
-                style: theme.textTheme.headlineLarge,
-              ),
-            ],
+      child: Stack(
+        children: [
+          const Align(child: Circles()),
+          Align(
+            child: TextArea(query: query),
+          ),
+        ],
+      ),
+    );
+  }
+}
+
+class TextArea extends StatelessWidget {
+  @visibleForTesting
+  const TextArea({required this.query, super.key});
+
+  final String query;
+
+  @override
+  Widget build(BuildContext context) {
+    final l10n = context.l10n;
+
+    return Column(
+      mainAxisAlignment: MainAxisAlignment.center,
+      children: [
+        Text(
+          l10n.thinkingHeadline,
+          textAlign: TextAlign.center,
+          style:
+              VertexTextStyles.body.copyWith(color: VertexColors.flutterNavy),
+        ),
+        const SizedBox(
+          height: 40,
+        ),
+        Padding(
+          padding: const EdgeInsets.symmetric(horizontal: 300),
+          child: Text(
+            query,
+            textAlign: TextAlign.center,
+            style: VertexTextStyles.display,
           ),
         ),
-      ),
+      ],
+    );
+  }
+}
+
+class Circles extends StatelessWidget {
+  const Circles({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    const backgroundColor = Colors.transparent;
+    const borderColor = VertexColors.googleBlue;
+    return const Stack(
+      children: [
+        Circle(
+          dotted: true,
+          backgroundColor: backgroundColor,
+          borderColor: borderColor,
+          radius: 162,
+        ),
+        Circle(
+          dotted: true,
+          backgroundColor: backgroundColor,
+          borderColor: borderColor,
+          radius: 353,
+        ),
+        Circle(
+          dotted: true,
+          backgroundColor: backgroundColor,
+          borderColor: borderColor,
+          radius: 590,
+        ),
+      ],
     );
   }
 }
